+++ CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Create repository. [#1](https://github.com/policy-design-lab/pdl-api/issues/1)
+- Statecode endpoint in the endpoint. [#15](https://github.com/policy-design-lab/pdl-api/issues/15)
+- Created statecode table in database. [#16](https://github.com/policy-design-lab/pdl-api/issues/16)